+++ MailDaemon/functions/Install-MDDaemon.ps1
@@ -30,6 +30,9 @@
 	
 	.PARAMETER SentPath
 		The folder in which emails that were successfully sent are stored for a specified time before being deleted.
+
+	.PARAMETER FailedPath
+		The path where mails that could repeatedly not be sent are moved to.
 	
 	.PARAMETER DaemonUser
 		The user to grant permissions needed to function as the Daemon account.
@@ -41,6 +44,12 @@
 	
 	.PARAMETER MailSentRetention
 		The time to keep successfully sent emails around.
+
+	.PARAMETER MailAbandonThreshold
+		How long we attempt to send an email before abandoning it and moving it to -FailedPath.
+
+	.PARAMETER MailFailedRetention
+		How long we keep an abandoned email around before removing it entirely.
 	
 	.PARAMETER SmtpServer
 		The mailserver to use for sending emails.
@@ -58,6 +67,10 @@
 	
 	.PARAMETER UseSSL
 		Use SSL for sending emails.
+
+	.PARAMETER NoLogging
+		Disables logging.
+		Unless specified, this setup step will also prepare the windows eventlog by creating a dedicated eventlog for MailDaemon.
 
 	.EXAMPLE
 		PS C:\> Install-MDDaemon -ComputerName DC1, DC2, DC3 -TaskUser $cred -DaemonUser "DOMAIN\MailDaemon" -SmtpServer 'mail.domain.org' -SenderDefault 'daemon@domain.org' -RecipientDefault 'helpdesk-t2@domain.org'
@@ -85,6 +98,9 @@
 		
 		[string]
 		$SentPath,
+
+		[string]
+		$FailedPath,
 		
 		[string]
 		$DaemonUser,
@@ -94,6 +110,12 @@
 		
 		[Timespan]
 		$MailSentRetention,
+
+		[Timespan]
+		$MailAbandonThreshold,
+		
+		[Timespan]
+		$MailFailedRetention,
 		
 		[string]
 		$SmtpServer,
@@ -108,7 +130,10 @@
 		$RecipientDefault,
 
 		[switch]
-		$UseSSL
+		$UseSSL,
+
+		[switch]
+		$NoLogging
 	)
 	
 	begin {
@@ -161,6 +186,7 @@
 		#region Setup Task Configuration
 		if (-not $NoTask) {
 			$action = New-ScheduledTaskAction -Execute powershell.exe -Argument "-NoProfile -Command Invoke-MDDaemon"
+			if ($NoLogging) { $action = New-ScheduledTaskAction -Execute powershell.exe -Argument "-NoProfile -Command Invoke-MDDaemon -NoLogging" }
 			$triggers = @()
 			$triggers += New-ScheduledTaskTrigger -AtStartup -RandomDelay "00:15:00"
 			$triggers += New-ScheduledTaskTrigger -At "00:00:00" -Daily
@@ -184,7 +210,7 @@
 		#endregion Setup Task Configuration
 		
 		#region Preparing Parameters
-		$parameters = $PSBoundParameters | ConvertTo-PSFHashtable -Include 'PickupPath', 'SentPath', 'MailSentRetention', 'SmtpServer', 'SenderDefault', 'RecipientDefault', 'UseSSL'
+		$parameters = $PSBoundParameters | ConvertTo-PSFHashtable -Include 'PickupPath', 'SentPath', 'FailedPath', 'MailSentRetention', 'MailAbandonThreshold', 'MailFailedRetention', 'SmtpServer', 'SenderDefault', 'RecipientDefault', 'UseSSL'
 		
 		$paramMainInstallCall = @{
 			ArgumentList = $parameters
@@ -246,21 +272,35 @@
 			if ($TaskUser) { $parametersSave['AccessAccount'] = $TaskUser }
 			Save-MDCredential @parametersSave
 			
-			$parametersInvoke = @{ $parametersInvoke['ComputerName'] = $ComputerName }
+			$parametersInvoke = @{ ComputerName = $ComputerName }
+			if ($Credential) { $parametersInvoke['Credential'] = $Credential }
 			Invoke-PSFCommand @parametersInvoke -ScriptBlock {
 				Set-MDDaemon -SenderCredentialPath "C:\ProgramData\PowerShell\MailDaemon\senderCredentials.clixml"
 			}
 		}
 		#endregion Securely store credentials
+
+		#region Setup Logging
+		if (-not $NoLogging) {
+			Invoke-PSFCommand @parametersInvoke -ScriptBlock {
+				if ($PSVersionTable.PSVersion.Major -gt 5 -and -not $IsWindows) { return }
+				Set-PSFLoggingProvider -Name eventlog -InstanceName MailDaemonInvoke -LogName MailDaemon -Source MailDaemon -Enabled $true -Wait
+				Write-PSFMessage -Message "Setting up MailDaemon logging"
+				Disable-PSFLoggingProvider -Name eventlog -InstanceName MailDaemonInvoke
+			}
+		}
+		#endregion Setup Logging
 		
 		#region Setup Task
 		if (-not $NoTask) {
-			foreach ($computerObject in $ComputerName) {
-				if ($ComputerName.Type -like 'CimSession') { $parametersRegister["CimSession"] = $computerObject.InputObject }
-				elseif (-not $ComputerName.IsLocalhost) { $parametersRegister["CimSession"] = $ComputerName }
-				
-				$null = Register-ScheduledTask @parametersRegister
-			}
+			Invoke-PSFCommand @parametersInvoke -ScriptBlock {
+				param ($ParametersRegister)
+
+				$taskObject = Get-ScheduledTask -TaskName $ParametersRegister.TaskName -ErrorAction Ignore
+				if ($taskObject) { $taskObject | Unregister-ScheduledTask }
+
+				$null = Register-ScheduledTask @ParametersRegister
+			} -ArgumentList $parametersRegister
 		}
 		#endregion Setup Task
 	}